# doc/api.yaml
openapi: 3.0.1
info:
  title: WASAPhoto APIs
  version: 1.0.3
  description: >
    **"Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto!"**
    
    WASAPhoto is web application based on RESTfull architecture to allows upload your photos directly from your PC, and they will be visible to everyone following you.
  license:
    name: GNU General Public License v3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
  
  contact:
    name: Andrea Pizzi
    url: https://github.com/pizzi-andrea/WASAPhoto
    email: pizzi.1995517@studenti.uniroma1.it
servers:
  - url: http://localhost:3000
paths:
...
  /users/{uid}/myStream/:
    summary: user stream photos
    description: >
      this resource rappresent stream of photos associated a user.
      the photos in the stream are taken by followed users. The user
      can put like or comment on photos in the stream.
    
    parameters:
        - $ref: "#/components/parameters/Uid"

    get:
      tags:
        - photo
        - stream
      summary: list personal stream photos 
      description: >
        given *uid* of user that who wants to get all photo in his stream
      
      parameters:
        - $ref: "#/components/parameters/QueryLimit"
        - $ref: "#/components/parameters/UsernameFilter"
        
        - name: sortBy
          in: query
          required: false
          schema:
            description: >
              options for sorting photos
            type: string
            enum: ["dataUpdate"]
            default: "dataUpdate"
          description: >
            ordering photo getted by value
       
        - name: reverse
          in: query
          required: false
          schema:
            description: >
              type value ordering flag
            type: string
            enum: ["asc", "desc"]
            default: "desc"
            
          description: >
            order photo in reverse order
        
      security:
        - BearerAuth: []
      operationId: getMyStream
      
      responses:
        "200":
          description: >
            success, getted stream photos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamPhotos"
        "404":
          $ref: "#/components/responses/UserNotFound"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
            
        "204":
          description: >
            empty body, found nothing
        "500":
          $ref: "#/components/responses/ServerError"
  /users/{uid}/myStream/{photoId}:
  
    parameters:
        - $ref: "#/components/parameters/Uid"
        - $ref: "#/components/parameters/PhotoId"
    
    get:
      
      tags:
        - photo
        - stream
        - user
      
      security:
        - BearerAuth: []
      
      summary: get photo from stream
      description: >
        give *uid* and *photoId* and get photo associated
        
      operationId: getPhotoMyStream
      responses:
        "200":
          description: >
            success, getted desired photo from stream
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
          
        "404":
          $ref: "#/components/responses/UserNotFound"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        
        "204":
          description: >
            empty body, found nothing
        "500":
          $ref: "#/components/responses/ServerError"
...
  /users/{uid}/myPhotos/{photoId}/: #
    summary: photo updated by user
    description: >
      this resource rappresent the streams of photos update by user in time
  
    parameters:
      - $ref: "#/components/parameters/Uid"
      - $ref: "#/components/parameters/PhotoId"
          
    delete:
      tags:
        - photo
      summary: delete photo updated
      description: >
        provide uid and photoID and delete the photo corresponding to photoID
    
      operationId: deletePhoto
      security:
        - BearerAuth: []
        
      responses:
        "204":
          description: >
            success, deleted photo
        "404":
          $ref: "#/components/responses/UserNotFound"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        "500":
          $ref: "#/components/responses/ServerError"
    
    get:
      tags:
        - photo
      security:
        - BearerAuth: []
      summary: get photo
      description: >
        give *uid* and photo id and get photo with photo id gived
      operationId: getPhoto
      
      responses:
        "200": 
          description: >
            success, photo founded
          
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Photo"
          
        "404":
          $ref: "#/components/responses/UserNotFound"
      
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        
        "204":
          description: >
            empty body, found nothing
        "500":
          $ref: "#/components/responses/ServerError"
...
  /users/{uid}/myPhotos/{photoId}/comments/:
    
    parameters:
      - $ref: "#/components/parameters/Uid"
      - $ref: "#/components/parameters/PhotoId"
    
    post:
      tags:
        - comment
        - photo
      summary: add comment a photo
      description: >
        give user id and photo id if user is owner of photo add a comment
        stored in the body of request
      
      
      
      requestBody:
        description: > 
          comment to add
        content:
          application/json:
            schema: 
              $ref: "#/components/schemas/Comment"
      
      security:
        - BearerAuth: []
      operationId: commentPhoto
      
      responses:
        "201":
          description: >
            success, new comment added to photo
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Comment"
          
        "404":
          $ref: "#/components/responses/UserNotFound"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        "500":
          $ref: "#/components/responses/ServerError"
    
    get:
      security:
        - BearerAuth: []
        
      tags:
        - photo
        - comment
      summary: get comments on photo
      description: >
        give photo id and user id of owner photo then get all comments on photo
      
      parameters:
        - $ref: "#/components/parameters/QueryLimit"
        - $ref: "#/components/parameters/UsernameFilter"
        - name: sortByData
          in: query
          required: false
          schema:
            description: >
              flag for sorting comments on photo by update data
            type: boolean
            default: true
          description: >
            sort comments by data creation
        
      
      operationId: getComments
      responses:
        "200":
          description: >
            success, get all comments on photo
          
          content:
            application/json:
              schema:
                description: >
                  list of comments on photo
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
                minItems: 1
                maxItems: 500

        "404":
          $ref: "#/components/responses/UserNotFound"
        "204":
          description: >
            empty body, found nothing
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        "500":
          $ref: "#/components/responses/ServerError"
  /users/{uid}/myPhotos/{photoId}/comments/{commentId}:
    summary: comment on photo
    description: >
      this resource rappresent a single comment on photo
    
    parameters:
      - $ref: "#/components/parameters/Uid"
      - $ref: "#/components/parameters/PhotoId"
      - $ref: "#/components/parameters/CommentId"
    delete:
      tags:
        - comment
        - photo
      summary: delete comment on photo
      description: >
        give user id of owner photo, photo id and remove comment on photo identified from comment id given
      
      security:
        - BearerAuth: []
        
      operationId: uncommentPhoto
      responses:
        "204":
          description: >
            success, deleted comment
        "404":
          $ref: "#/components/responses/UserNotFound"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        "500":
          $ref: "#/components/responses/ServerError"
    
    get:
      tags:
        - comment
        - photo
      
      security:
        - BearerAuth: []
      
      summary: get comment on photo
      description: >
        give user id photo owner, photo id and id of comment on photo and get it
      
      operationId: getComment
      
      responses:
        "200":
          description: >
            get comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "404":
          description: >
            user or comment not found
        
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        
        "204":
          description: >
            empty body, found nothing
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  parameters:
    QueryLimit:
      description: >
        security parameter to limit the number of items in a response and avoid DOS attacks
      name: limit
      in: query
      required: false
      schema:
        description: >
          format limit values
        type: integer
        format: int
        minimum: 1
        example: 50
    UsernameFilter:
      description: >
        research in list of users by username 
      name: username
      in: query
      required: false
      schema:
        type: string
        pattern: '^.*?$'
        minLength: 1
        maxLength: 16
        example: "marioross"
    Uid:
      name: uid
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Id"
      description: >
        user id
    PhotoId:
      name: photoId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Id"
      description: >
        photo id
    CommentId:
      name: commentId
      in: path
      required: true
      description: >
        comment id
      schema:
        $ref: "#/components/schemas/Id"
...
  responses:
    UnauthorizedError:
      description: >
        Access token is missing or invalid 
    UnauthorizedToken:
      description: >
        the submitted token does not have the necessary permissions to access this resource
    ServerError:
      description: >
        server internal error, for more information see log file
    UserNotFound:
      description: >
        Not found, user not found
...
  securitySchemes:
    
    BearerAuth:
      description: >
        basic safety mechanism based on "bearer" mechanism. APIs that require a minimum level of security can adopt this method
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      description: >
        this object rappresent a user
      type: object
      properties:
        uid:
          $ref: "#/components/schemas/Id"
        username: 
          $ref: "#/components/schemas/Username"
      required:
        - uid
        - username
...
    Username:
      description: >
        username of a user
      type: string
      pattern: '^.*?$'
      minLength: 3
      maxLength: 16
      example: "marioross"
    Id:
      description: >
        Identificator at 64-bit
      type: integer
      format: int64 
      readOnly: true
      example: 4570 
...
    Photo:
      description: >
        this object rappresent a photo
      type: object
      properties:
        photoId:
          $ref: "#/components/schemas/Id"
        timeUpdate:
          $ref: "#/components/schemas/TimeStamp"

        imageData:
          description: >
            data
          type: string
          format: binary
          minLength: 0
          maxLength: 5000000
          

        
        descriptionImg:
          description: >
            image description
          type: string
          example: "a good photo!"
          minLength: 1
          maxLength: 250
          pattern: "^.*?$"
      required:
        - imageData
        - timeUpdate
    StreamPhotos:
      description: >
        model of stream of photos
      type: array
      items:
        $ref: "#/components/schemas/Photo"
      minItems: 0
      maxItems: 100
    TimeStamp:
      description: >
        this components describe timestamp value
        conform to [RFC3339](https://datatracker.ietf.org/doc/html/rfc3339#section-5.6) specification
      type: string
      format: data-time
      pattern: '^[1-9]\d{3}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}Z$'
      example: "2017-07-21T17:32:28Z"
      minLength: 20
      maxLength: 21
      readOnly: true
    Comment:
      description: > 
        this object rappresent a comment on a photo.
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        commentId:
          $ref: "#/components/schemas/Id"
        text:
          description: >
            comment text encoded in UNICODE format
          type: string
          minLength: 1
          maxLength: 250
          pattern: "^.*$"
          example: 
            "😀 i like you photo! 😀"
        timeStamp:
          $ref: "#/components/schemas/TimeStamp"
      required:
        - user
        - text
        - timeStamp
        - commentId
...